Refactored the Inventory System, Fixed a bug with PermHP, Fixed access violation error, Code clean up.
Changed the Inventory system to rely on a vector of a pair. This also allows it to be saved into the XML more easily (which is now a thing).  Fixed a bug with Permanent HP items counting towards HP Potions quantity. Fixed a bug with drawing inventory quantity. Cleaned up some obsolete code.