Fixed (level) dropped items memory leak
Dropped items are not created as new each time. The specific item is created once and each time the item is dropped by the enemy, it has it's own handling per index in the vector. This reduces memory drastically as memory will never go up now for each time an item that has been created before will not be re-created which adds to the reference level objects data.