Refactored Projectiles class / Added completed cutscenes to the save files / Bug fixes
Refactored the Projectile class to be much more smoother and efficient. This also gets rid of the nasty exception shooting bullets in different directions at the same time gave. The refactoring also allows for new types of bullets to be created with ease.
Cutscenes that are completed are now added to the save file and do not get added once you leave the map.
Fixed some bugs with Shade