Added player status effects / Added collisions for environment hazards
The player can now collide with hazards that the level environment poses.  Such as Lava or Poison plants.
The player also receives a status effect of "POISONED" when they collide with a poison plant. This is a DOT based attack from the level. Every 1 second take damage for 8 seconds.
Lava is a constant HP loss tick.