Added iFrame animations / Added BGM & Sound effects for multiple things / Improved status effect text
Added an animation for iFrames (It's just setting the current animation to invisible and the next animation will show).
Added BGM to every map and loading a map that has the same BGM will keep playing the current BGM at the current timestamp.
Added sound effects to the player and enemies.
Generalized the status effect text drawing function.