Updated getTilesetPosition() formula
Fixed a calculation error when determining the Tileset position.
The modular division for the 16th column will be equal to 0 for tsxx, and then subtracted by 1 which then puts the position at a negative number when multiplied by the width. Same with the height as well. So that would mean -16,-16.
This corrected mathematical formula will ensure that this does not happen.