Added Event messages & Battle messages
Event messages will show text based on the interaction the player has made. I.E: If a player saves, it will display a saved successfully message. Or if a player tries to open a locked portal, it will display Key required.
Battle messages will appear on the bottom right of the screen and will display useful combat stats such as EXP gained and what enemy was killed along with the total amount of kills on that enemy.