Improved the cameras vertical scrolling
When there are small changes in the players Y coordinates such as traversing a slope, the camera would flicker as slopes tend to push the player up to keep them above the slope and down (from gravity). Using checkInRange() to see if there was such change, the camera will use the previous Y to mitigate any flickering.