Changed corruption damage multiplier formula
Corruption has a new varying formula that is still RNG based. However, the gain is much lower than it before. The player will receive a random value between 0.30 and 0.80 that will be added to their damage. The higher a player's soul strength, the greater the chance of getting a high roll.