Updated Exp system
Gaining exp from enemies is now on a separate system than gaining exp elsewhere (I.E: quests). There is a standard -0.25 penalty after the player reaches level 1. And if the enemy killed is a lower level than the player, there are further penalties applied to the exp gained by the player.