Added Hexes that can be applied to the player
Hexes casted onto the player will be different depending on the type of hex that was used. For now, there is just 1 hex that will reduce the players final damage. The next update will consist of handling that reduction which will be by the means of introducing a final damage calculation that will account for the players damage mod and damage reduction to create a final damage value.