Implemented a healing mechanic for Shade
The Shade can now heal after reaching a 35% HP threshold. There is a initial heal of +50 HP. Then every 5 seconds it will heal +15 HP until it has done so 5 times. After that, it's ability to heal will be on cooldown for 30 seconds and can only be triggered again once the 35% HP threshold criteria is met.