Fixed a bug (and possible exploit) with pausing the game
The elapsed time when resuming a game after pausing would receive a very large input depending on how long the game was paused for. This was due to the LAST_UPDATE_TIME not being updated after the game was resumed. We set it to the current ticks - 1 to account for the difference and to allow for a smooth and consistent transition for the next iteration to use in it's update/drawing.