Refactor Inventory class
Inventory class will have two maps. One that stores the ItemID and what Item has that ID in the prototype map and another map that is the players current inventory where items will be added by using it's ItemID. It will also call the Items overridden use() function when used.