Implement drawMultiLine()
Takes in standard parameters to draw text but returns an int value that is equal to the height of the textbox. Using the return value, the next text drawing(s) will be evenly spaced out by 1px. This will handle multiple sentences more effectively instead of random spacing in between.